Dia2/gemini-deploy/requirements.txt
#Flask==2.0.2
#numpy==1.21.1
#pandas==1.3.1
#requests==2.26.0
#scikit-learn== 1.4.2
#streamlit==1.0.0
#waitress==2.1.0
#fastapi==0.115.6
#google-generativeai==0.8.3

numpy
pandas
requests
gunicorn
langchain 
langchain-google-genai 
google-generativeai 
python-dotenv
fastapi
google-generativeai
fastapi[standard]
